feat: create a layout component for home and thank-you page
