+++ src/constants/thankInfo.js
@@ -1,4 +1,5 @@
 export const ThankInfo = {
+  page: 'thankyou',
   speechBubble: [
     {
       id: 0,
@@ -23,6 +24,7 @@ export const ThankInfo = {
   ],
   bottomText: '我們希望能透過這個體驗讓您更加了解 Scrum 唷！ 謝謝您撥了點時間來體驗這整個 Scrum 的流程，希望對您們未來實務上能有所幫助！！',
   button: '再體驗一次！',
+  next: '/',
 };
 
 export default ThankInfo;
